chore: organized the endpoints better.
- Reorganized the files.
- The sorter of the swagger tags is now working.
- Added new tags, and Updated a few.
