+++ Endpoints/Match/DeleteMatchEndpoints.cs
@@ -1,0 +1,11 @@
+using Microsoft.AspNetCore.Builder;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.AspNetCore.Mvc;
+
+public static class DeleteMatchEndpoints
+{
+    public static void MapDeleteMatchEndpoints(this WebApplication app)
+    {        
+    }
+}
